refactored to use a shared input field
